for some reason when we sell stuff, it displays our gold all wrong, unless we cast it to an int in the printf

+++ src/misc3.c
@@ -286,7 +286,7 @@ prt_lnum (header, num, row, column)
 {
   vtype out_val;
 
-  snprintf (out_val, sizeof (out_val), "%s: %6ld", header, num);
+  snprintf (out_val, sizeof (out_val), "%s: %6ld", header, (int)num);
   put_buffer (out_val, row, column);
 }
